fix, added urls to test and removed urls.ts
