+++ packages/neon-domain/__integration__/NeoNS.ts
@@ -1,9 +1,33 @@
 import * as neonCore from "@cityofzion/neon-core";
 import domainPlugin from "../src/index";
-import { getMainnetUrls } from "./urls";
 
 const neonJs = domainPlugin(neonCore);
 const { CONST, rpc, u, api, wallet } = neonJs;
+
+export function getMainnetUrls(): string[] {
+    return [
+        "http://seed1.ngd.network:10332",
+        "http://seed2.ngd.network:10332",
+        "http://seed3.ngd.network:10332",
+        "http://seed4.ngd.network:10332",
+        "http://seed5.ngd.network:10332",
+        "http://seed6.ngd.network:10332",
+        "http://seed7.ngd.network:10332",
+        "http://seed8.ngd.network:10332",
+        "http://seed9.ngd.network:10332",
+        "http://seed10.ngd.network:10332",
+        "https://seed0.cityofzion.io:443",
+        "https://seed1.cityofzion.io:443",
+        "https://seed2.cityofzion.io:443",
+        "https://seed3.cityofzion.io:443",
+        "https://seed4.cityofzion.io:443",
+        "https://seed5.cityofzion.io:443",
+        "https://seed6.cityofzion.io:443",
+        "https://seed7.cityofzion.io:443",
+        "https://seed8.cityofzion.io:443",
+        "https://seed9.cityofzion.io:443"
+    ]
+};
 
 let MAINNET_URL = "";
 let provider;
